# server/swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getRunning","summary":"Simple get endpoint","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseApiDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseApiDto"}}}}}}},"/api/tarefa":{"get":{"operationId":"TarefaController_getAll","summary":"Lista todas as tarefas","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"result":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/TarefaResponseDto"}}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExceptionResponseApiDto"}}}}},"tags":["Tarefas"]},"post":{"operationId":"TarefaController_createBadge","summary":"Criar tarefa","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CriarTarefaRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"result":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/TarefaResponseDto"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExceptionResponseApiDto"}}}}},"tags":["Tarefas"]}},"/api/tarefa/{tarefaId}":{"get":{"operationId":"TarefaController_getById","summary":"Obter tarefa por id","parameters":[{"name":"tarefaId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"result":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/TarefaResponseDto"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExceptionResponseApiDto"}}}}},"tags":["Tarefas"]},"put":{"operationId":"TarefaController_updateBadge","summary":"Atualizar tarefa","parameters":[{"name":"tarefaId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AtualizarTarefaRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"result":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/TarefaResponseDto"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExceptionResponseApiDto"}}}}},"tags":["Tarefas"]}}},"info":{"title":"poc-nestjs-mongoose","description":"proof of concept","version":"v1","contact":{}},"tags":[{"name":"api","description":""}],"servers":[],"components":{"schemas":{"ResponseApiDto":{"type":"object","properties":{"result":{"type":"boolean","description":"Result of response","example":true},"message":{"type":"string","description":"Message of result","example":"Executado com sucesso!"}},"required":["result","message"]},"DataResponseApiDto":{"type":"object","properties":{"result":{"type":"boolean","description":"Result of response","example":true},"message":{"type":"string","description":"Message of result","example":"Executado com sucesso!"},"data":{"type":"object","description":"Data of result"}},"required":["result","message","data"]},"TarefaResponseDto":{"type":"object","properties":{"descricao":{"type":"string","description":"Descricao da tarefa"},"concluido":{"type":"boolean","description":"Tarefa concluida"}},"required":["descricao","concluido"]},"ExceptionResponseApiDto":{"type":"object","properties":{"result":{"type":"boolean","description":"Result of response","example":false},"message":{"type":"string","description":"Message of result","example":"BadRequest"}},"required":["result","message"]},"CriarTarefaRequestDto":{"type":"object","properties":{"descricao":{"type":"string","description":"Descrição da tarefa","example":""},"concluido":{"type":"boolean","description":"Tarefa concluida","example":false}},"required":["descricao","concluido"]},"AtualizarTarefaRequestDto":{"type":"object","properties":{"descricao":{"type":"string","description":"Descrição da tarefa","example":""},"concluido":{"type":"boolean","description":"Tarefa concluida","example":false}},"required":["descricao","concluido"]}}}}